Partial setup step 1
Added the HTML setup for the south row and some basic styling for the properties that will be required going forward.

Some placeholder backgrounds for the image-based properties have been created (stations, chance, community chest)

Popup created with a close function on an eventListener. This will get heavy use throughout the game.

A very preliminary version of the chance/community chest drawing function. This currently puts the content of the card in the popup window but doesn't do anything else right now. At present picks at random but this needs to be shuffled and stacked in the future.