Token moving (preliminary)
A very basic token moves around the board based on the dice roll, although it only works once around the board before it breaks. The code behind this requires much improvement, including a messy switch statement for determining the position, and SCSS that I'm sure could be improved.

Despite its flaws, I'm committing this because it does work.