Player turns preliminary
Fixed broken dice rolling following adding multiple tokens. The game now adds a class to the body element which keeps a record of what player's turn it is, and the dice rolling functions now correctly determine which token they should be moving. However, there is not yet functionality to end a player's turn - it is currently stuck on player 1's turn.