Get out of jail - pay and doubles
Up until now sending players to jail just meant moving them to position 10. No more. The game now stores whether the player is in jail, and currently gives them the option of 50 or attempting to roll doubles to get out.

According to official Monopoly rules the player is obligated to pay to get out or use a 'Get out of jail free' card after 3 failed doubles rolls. The game does not currently do this.