Currency Symbol
Added a currency symbol in front of all the money values.

This doesn't sound like it should have been very complicated but it was.  Since 2008 Monopoly uses a currency called 'Monopoly dollars'. The symbol for this is not an official character with an alt code and does not appear in any fonts. However, it is effectively an upside  down '₩' character - this is for 'Won' - the currency used in North and South Korea.

Therefore it was necessary to insert this in a span and rotate it. It also required a bit of font size tweaking and positioning to get it to sit where it looks natural alongside the numbers.