updatePlayerDetails overhaul
updatePlayerDetails() needed rewriting after the game was updated to use currency symbols and simply comparing values didn't work any more.

At that point, I also realised that looping through all of the values was somewhat inefficient. There are several properties that the player never sees (like the inJail), and of the others there are some that are never going to change (like the player name).

Therefore it now seems like it's more efficient to just update the properties which are likely to have actually changed.