Tokens
Players now get a token rather than just a coloured circle. Overhaul of positionToken() to get it to position the tokens in the middle of the property, and stack them if there are multiples. There is an issue in the way this works; it's unlikely to be triggered in normal gameplay but I have officially reported it to myself and I will revisit at a later date.

The beginnings of the token selector have been scrapped - I plan on doing this a different way.