Even building with bank check
The new code which forces players to build evenly now also prevents players from building when there are no houses/hotels left in the bank. Button text is updated as appropriate.

I also rejigged the tokens array. While the dog is the most iconic piece, the thimble is ultimately the best ;)

Putting the thimble second in the array makes it the default piece for player 2.