Auctions
When landing on an unowned property, a player may choose to send the property to auction. They are prevented from closing the popup window - they are forced to either buy or send to auction.

The displayPropertyDetails() function has been split into two - the bit that generates the HTML and the bit that makes it appear in the popup. For the purposes of the auction screen I needed the details of the property without it being forced into the popup.