General tidying
Updated the readme file to contain details of all of the new functions from this branch - newGameDiceRoll() and all of its private functions, plus createConfetti which I think is going to be useful down the line.

Also tidied up the SCSS so there is now a section for general purpose animations.