Trade money validity check
Ensured that players had to enter a value for the money that was between the min and max parameters of the input (which is set based on how much money they have).

This was necessary as while there are min and max parameters set and the arrows on the input prevent them going above/below this, the browser does not prevent them from typing in something invalid.

The game now also disables the propose trade button if there is invalid input.