Trade window property details
When hovering over properties in the trade negotiation window, a small window will appear to show the title deed of that property.

The code checks whether the screen is large enough to display this info, and will disable the windows if it's not. While this isn't ideal, this game isn't intended to be played on smaller screens, plus this information isn't essential, it's just a nice bonus.