Fake dice roll functional and visual improvements
Created the buttons for fake dice rolls 1-12 in the testing panel, as well as some nice visual improvements.

While the testing panel was originally created for me to test things with while the game is in development, I may keep it in the end game (although make it harder to get to). Therefore I would like it to look good.