Quickstart mode improvements
Updated the quickstart mode so that players automatically own properties, some of which have houses/hotels, some are mortgaged, get out of free cards are owned etc etc. Utilities and stations are owned as groups.

The way it's implemented is a little crude, but it should make testing significantly easier.