Back button tweaks
Moved the back button into its proper position in the property overview, and also made it so that you can use backspace on the keyboard as well.

Some refactoring of fullPortfolioView() was required as it will no longer necessarily be a click event that triggers it, so the event object won't have the properties it was previously relying on.