Go back past Go feed message
The feed will now display a correct message if a 'go back' card causes the token to move back past Go.

While on a standard set I don't believe it's possible to go back past Go with a card, it's worth including in case any alternative boards I may add in the future have this possibility.