Player summary refactoring
Refactored the player summary generation to have the player id as a separate attribute rather than part of the ID. Admittedly this is how I should have coded it in the first place.

While this seems off topic for the subject of this branch, this has repercussions for how I remove players from the game.