Transaction queue
Added a new transactionQueue variable for storing a queue of transactions, and refactored the payMoney() function to add to this when receiving an object. When either bankruptcy proceedings or payMoney concludes (depending on whether the debtor has enough money), it will check if there are more transactions in the queue then start again.