Tooltips
The beginnings of code for tooltips. This took a lot more refactoring than I thought it would, and I fear I have opened up a can of worms by doing this.

Tooltips are in my stretch goals, but there is one thing in my main goals which can also use the same code, so I've done this now.