Auction bid tooltip
The money inputs in the auction bidding screen now display a tooltip if the player is attempting to bid more money than they have.

According to the rules you are allowed to bid more money than you have, but you will be expected to raise the money through trades, mortgages or selling buildings afterwards.