Player colour temporary fix
Eventually the plan is to have players be able to pick their own colours. I was previously randomly generating them but black text wasn't showing up on top of the darker colours. As a stretch goal I would like to make it so that the game detects this and changes the text and icons on top to white, but for now I have updated the way the random colours are generated to only generate lighter colours for which black text works.