gameState
Created a variable called gameState and a function called initialiseGameState() to store info about each space. This is the first step in a major refactoring where information on ownership, mortgages and buildings are all stored in this central location rather than shared between the spaces/players arrays. This is to pave the way to be able to save your game and come back to it later.